+++ cross-media-measurement/src/main/kotlin/org/wfanet/measurement/gcloud/spanner/BUILD.bazel
@@ -19,6 +19,7 @@ kt_jvm_library(
         "//imports/java/picocli",
         "//imports/kotlin/kotlinx/coroutines:core",
         "//src/main/kotlin/org/wfanet/measurement/common",
+        "//src/main/kotlin/org/wfanet/measurement/common/identity",
         "//src/main/kotlin/org/wfanet/measurement/gcloud/common",
     ],
 )
