+++ cross-media-measurement/src/main/kotlin/org/wfanet/measurement/gcloud/spanner/BaseSpannerReader.kt
@@ -1,4 +1,4 @@
-// Copyright 2020 The Cross-Media Measurement Authors
+// Copyright 2020 The Measurement System Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,14 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.kingdom.deploy.gcloud.spanner.readers
+package org.wfanet.measurement.gcloud.spanner
 
 import com.google.cloud.spanner.Statement
 import com.google.cloud.spanner.Struct
 import java.util.logging.Logger
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import org.wfanet.measurement.gcloud.spanner.AsyncDatabaseClient
 
 /** Base abstraction for reading from Spanner. */
 abstract class BaseSpannerReader<T> {
