+++ cross-media-measurement/src/main/kotlin/org/wfanet/measurement/kingdom/deploy/gcloud/spanner/queries/StreamReports.kt
@@ -16,13 +16,14 @@ package org.wfanet.measurement.kingdom.deploy.gcloud.spanner.queries
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
+import org.wfanet.measurement.gcloud.spanner.BaseSpannerReader
+import org.wfanet.measurement.gcloud.spanner.SpannerQuery
 import org.wfanet.measurement.gcloud.spanner.appendClause
 import org.wfanet.measurement.internal.kingdom.Report
 import org.wfanet.measurement.kingdom.db.StreamReportsFilter
 import org.wfanet.measurement.kingdom.db.hasStateFilter
 import org.wfanet.measurement.kingdom.deploy.gcloud.spanner.common.StreamReportsFilterSqlConverter
 import org.wfanet.measurement.kingdom.deploy.gcloud.spanner.common.toSql
-import org.wfanet.measurement.kingdom.deploy.gcloud.spanner.readers.BaseSpannerReader
 import org.wfanet.measurement.kingdom.deploy.gcloud.spanner.readers.ReportReader
 
 /**
